Ajout alerte suppression ficheir

+++ src/MainBundle/Resources/public/js/ide.js
@@ -59,7 +59,7 @@ function requestAndSetLanguage(language){
       $("#select-modele").val(-1);
       console.log(modeles[0].model);
       $("#show-model").text(modeles[0].model).html();
-      $("#show-model").html($("#show-model").html().replace(/\n/g,"<br>"));
+      $("#show-modelhttps://trello.com/b/PCq4IDlC/lide-licence-informatic-development-environment").html($("#show-model").html().replace(/\n/g,"<br>"));
     },
   });
 }
@@ -142,8 +142,30 @@ function removeFile(id){
 }
 
 $("#rm_file").click(function (){
-  removeFile(currentFile);
+  if(confirm("Attention ! Toute donnée non sauvegardées seront perdu !")){
+    removeFile(currentFile);
+    
+  }
 });
+
+function saveFile(id){
+  if(id >= 0 && id < files.length){
+    if(id == currentFile){
+      files[currentFile].content = editor.getValue();
+    }
+    var blob = new Blob( [files[id].content] , {type: "text/plain;charset=utf-8"});
+    saveAs(blob, files[id].name);
+  }
+}
+
+function saveAllFile(){
+  for(i = 0; i < files.length; ++i){
+    saveFile(i);
+  }
+}
+
+$("#btn-save").click(saveAllFile);
+
 
 $("#btn-create-file").click(function (){
   var fileContent = "";
@@ -188,7 +210,6 @@ $("#console-toogle").click(function (){
 
   $("#block-editor").toggleClass("col-12");
   $("#block-editor").toggleClass("col-8");
-
 });
 
 //Selection d'un nouveau langage
